Pixel-font support
Added pixel-font support. This is messy and hacky and I dislike
it—<canvas> has no support for turning off text anti-aliasing, so it’s
hard to get the right effect. I’ll be reverting to a previous version
with a proper font, but I want to keep this work in case it comes in
useful later.