Made a bunch of changes
Yes, I know I ought not to make so many changes in a single commit, but
there you go. I got carried away.
Fixed a bug with Sketch when the Pokémon had not previously made a
move. Fixed a bug with Metronome and multi-turn moves. Added a load of
debug code to try to find and destroy the #121 bug. Sped up the `_`
function massively, with special cases. Tried to speed up other
functions by not setting function arguments within the functions
themselves.
Made some drawing objects only redraw when they need to, such as the
Textbox, which should, hypothetically, improve performance.
Added evolutions (some changes may be needed, such as fading the
evolution screen in and out, but the basic infrastructure is all
there), which fixes #62.
Added standard deviation to the FPS counter. Fixed a bug where health
values would not be integers while transitioning.