Added battle end callbacks
Now, when a battle ends, it will call the callback function provided,
supplying it with the outcome of the battle. This allows for far easier
integration with other systems and thus fixes #145. Additionally, this
has been taken advantage of to allow multiple battles to take place
consecutively in the example page without the user having to refresh
the page.