Added a move summary view
Now, when you hold shift while hovering over a move (either with the
mouse or on a selected move using the keyboard), you’ll see a number of
details regarding that move (including, perhaps most usefully, the PP).
This goes some way to adding #126, although for now makes use of an
even-more-coupled Input and Textbox (which will eventually be
addressed—see #155).