Added full Pokémon form(e)s
This resolves #89. Getting species-related properties is slower now, as
it has to go through an additional form(e) check, which probably does
not help #78. At some point I’ll have to try addressing that. Hopefully
I’ve caught all the properties that need to be checked this way, but I
wouldn’t be surprised to find a few that have slipped through the gaps
in the next few days/weeks/even longer. This is quite a major change
and opens up the opportunity for other major features, such as Mega
Evolutions (#24).