Modified the storage system to allow syncing with the server
Each trainer now has their own storage system. This is far more
versatile and more well-suited to situations with multiple non-NPC
trainers, such as those you’d find in an online game. This should
enable syncing the Pokémon sent to the PC with the server. We’re
getting very close to completing #149.