Fixed a bug with sending out the same Pokémon twice at once
I had actually added a `reserved` property to `battler`, but it wasn’t
being checked in all the right places. I also released that the
existing `switching` would work in its place (so it’s now used for both
Pokémon switching in and out), because Pokémon’s `battler`s are reset
as soon as they enter a battle.