Added the other battle actions to the delegates
This alone will not enable delegates to fully function (it only allows
the menu options to be enabled/disabled). It still needs to notify the
delegates when they are to perform an action. They also need to send
callback methods to the delegates (e.g. for when a new Pokémon is
needed to be switched in). (Though, technically, even at this stage,
they could simply call BattleContext’s methods directly, this is not a
clean solution).