Added input wellformedness validation
Hypothetically, this should prevent nasty people sending malformed
input via the battle system and causing crashes or other manners of
unpleasant things. This is the last feature necessary for the online
battle system to be completely secure and proof against meddlers. There
will no doubt be a host of bugs to fix, but the main framework is
there. Let the games begin. This closes #166.