Added a TrialServer object
This will be used to test out online battles without actually requiring
a proper server. The TrialServer simulates a server by sending the same
messages, but in the same page, making it far easier to debug. At the
moment, it only supports 1-person battles (so wild or NPC trainer, but
not multiplayer).

Also fixed a couple of issues with the server side input validation.