Fixed server-client synchronisation checks
This means as well as blocking invalid commands from players, we can
check if they’ve been mischievously trying to edit their own battles
locally (which won’t work of course, but it’s best not to encourage
such things). Synchronisation checks are not enforced yet (this might
be a good idea later). This closes #170. Also fixed a small issue with
Future Sight.